style: rename sidebar section title file
